+++ plugin/galerie/admin.php
@@ -45,7 +45,7 @@ switch ($action) {
             } else {
                 show::msg("Une erreur est survenue", 'error');
             }
-            header('location:index.php?p=galerie');
+            header('location:index.php?p=galerie&action=edit&id=' . $item->getId());
             die();
         }
         break;